# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stream API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh token",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Account is already exists"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get list of orders",
                "description": "Returns list of orders",
                "operationId": "getOrders",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "required": true,
                        "style": "deepObject",
                        "schema": {
                            "properties": {
                                "currency": {
                                    "type": "string"
                                },
                                "from": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            },
                            "type": "string",
                            "example": {
                                "currency": {
                                    "id": 1
                                },
                                "from": "2021-03-21T22:00:00.000+00:00"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Create order",
                "description": "Create new order",
                "operationId": "addOrder",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/newOrderData"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Address": {
                "title": "Address",
                "description": "Address model",
                "properties": {
                    "id": {
                        "description": "The unique identifier of a address.",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": "true"
                    },
                    "address": {
                        "description": "The full address.",
                        "type": "string",
                        "example": "Karlshof, 72351 Geislingen, Германия"
                    },
                    "lat": {
                        "description": "The latitude coordinate.",
                        "type": "number",
                        "format": "float",
                        "readOnly": "true"
                    },
                    "lng": {
                        "description": "The longitude coordinate.",
                        "type": "number",
                        "format": "float",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "Currency": {
                "title": "Currency",
                "description": "Currency model",
                "properties": {
                    "id": {
                        "description": "The unique identifier of a currency.",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": "true"
                    },
                    "name": {
                        "description": "The name of a currency.",
                        "type": "string",
                        "example": "USD"
                    },
                    "code": {
                        "description": "The code of a currency.",
                        "type": "string",
                        "example": "USD"
                    }
                },
                "type": "object"
            },
            "Customer": {
                "title": "Customer",
                "description": "Customer model",
                "properties": {
                    "base_id": {
                        "description": "The unique identifier of a customer.",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": "true"
                    },
                    "name": {
                        "description": "The name of a customer.",
                        "type": "string",
                        "example": "Liyah Swift"
                    },
                    "remote_id": {
                        "description": "The unique identifier of a customer in remote database.",
                        "type": "integer",
                        "format": "int64",
                        "example": "6554"
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
            "Market": {
                "title": "Market",
                "description": "Market model",
                "properties": {
                    "base_id": {
                        "description": "The unique identifier of a market.",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "description": "The market name.",
                        "type": "string",
                        "example": "comfyballs.no"
                    },
                    "remote_id": {
                        "description": "The unique identifier of a market in remote database.",
                        "type": "integer",
                        "format": "int64",
                        "example": "4"
                    },
                    "iconLink": {
                        "description": "The link for market icon.",
                        "type": "string",
                        "format": "url",
                        "example": "/images/markets/comfyballs.jpeg"
                    },
                    "color": {
                        "description": "The market color.",
                        "type": "string",
                        "example": "green"
                    }
                },
                "type": "object"
            },
...
            "newOrderData": {
                "title": "new Order Data",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Order"
                    },
                    {
                        "$ref": "#/components/schemas/Address"
                    },
                    {
                        "properties": {
                            "currency": {
                                "$ref": "#/components/schemas/Currency"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "customer": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "market": {
                                "$ref": "#/components/schemas/Market"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "status": {
                                "$ref": "#/components/schemas/OrderStatus"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "products": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/OrderProduct"
                                }
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "warehouse": {
                                "$ref": "#/components/schemas/Warehouse"
                            },
                            "id": {
                                "description": "The unique identifier of a order.",
                                "type": "integer",
                                "format": "int64",
                                "readOnly": "true"
                            },
                            "order_id": {
                                "description": "The unique identifier of a order in remote table.",
                                "type": "string",
                                "example": "9303"
                            },
                            "packing_cost": {
                                "description": "Packing cost of order",
                                "type": "number",
                                "format": "float",
                                "example": "123.33"
                            },
                            "created_at": {
                                "description": "Initial creation timestamp",
                                "type": "string",
                                "format": "date-time",
                                "readOnly": "true"
                            },
                            "updated_at": {
                                "description": "Last update timestamp",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "Order": {
                "title": "Order",
                "description": "Order model"
            },
            "OrderStatus": {
                "title": "OrderStatus",
                "description": "Order statuses",
                "properties": {
                    "base_id": {
                        "description": "The unique identifier of a order status.",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": "true"
                    },
                    "name": {
                        "description": "The name of a status.",
                        "type": "string",
                        "example": "new"
                    },
                    "remote_id": {
                        "description": "The unique identifier of a order status in remote table.",
                        "type": "integer",
                        "format": "int64",
                        "example": "4"
                    },
                    "color": {
                        "description": "The color of a market.",
                        "type": "string",
                        "example": "green",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Warehouse": {
                "title": "Warehouse",
                "description": "Warehouse model",
                "properties": {
                    "id": {
                        "description": "The unique identifier of a warehouse.",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "description": "The name of warehouse.",
                        "type": "string",
                        "example": "Vestby"
                    }
                },
                "type": "object"
            },
...
            "OrderProduct": {
                "title": "OrderProduct model",
                "description": "ClickHouse OrderProduct for store",
                "properties": {
                    "remote_id": {
                        "format": "int64",
                        "example": "301498"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": "738.95"
                    },
                    "profit": {
                        "type": "number",
                        "format": "float",
                        "example": "253.39"
                    },
                    "qty": {
                        "format": "int64",
                        "example": "2"
                    },
                    "name": {
                        "example": "TITAN LIFE Training Mat 180x60x1.5cm - Grey"
                    },
                    "link": {
                        "format": "url",
                        "example": "https://www.tights.no/butikk/titan-life-training-mat-180x60x1-5cm-grey/"
                    },
                    "image_link": {
                        "format": "url",
                        "example": "https://www.tights.no/wp-content/uploads/sites/7/2020/11/training_mat.jpg"
                    },
                    "discount": {
                        "type": "number",
                        "format": "float",
                        "example": "6.23"
                    },
                    "weight": {
                        "type": "number",
                        "format": "float",
                        "example": "2.05"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "token": {
                "type": "http",
                "description": "Bearer {JWT}",
                "name": "authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}